Squashed commit of the following:
commit 709b4826444e02d23f826e16c0474f06339e2196
Author: Dennis Adalbert Boghean <bogheanadalbertd@gmail.com>
Date:   Thu Jan 15 08:27:49 2026 +0100

    Get categorias

commit 60f0024560504a41e4c6d1168178b4edce67506c
Author: Dennis Adalbert Boghean <bogheanadalbertd@gmail.com>
Date:   Tue Jan 13 18:35:47 2026 +0100

    new VideoPrivateDTO.

commit 089cac019211a0a1db60c46ed746ce5e838cfe0f
Author: Dennis Adalbert Boghean <bogheanadalbertd@gmail.com>
Date:   Tue Jan 13 13:06:51 2026 +0100

    Servidor api, solo get videos

commit 79b7c32c8603db0144d68ae91150d5609317dae5
Author: Dennis Adalbert Boghean <bogheanadalbertd@gmail.com>
Date:   Tue Jan 13 13:06:14 2026 +0100

    Borrados archivos viejos

+++ backend/catalogo/src/main/java/blostesoftware/blostefix/catalogo/repositories/VideoRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import blostesoftware.blostefix.catalogo.models.Video;
 
-public interface VideoRepository extends JpaRepository<Video, Integer> {
-   
+public interface VideoRepository extends JpaRepository<Video, Long> {
 }
